correct testimonial modal
